+++ iter_test.go
@@ -1,3 +1,5 @@
+//go:build goexperiment.rangefunc
+
 package hi
 
 import (
